Implement certificate renewal

+++ daemon/certs.py
@@ -1,4 +1,5 @@
 import os
+import db
 
 
 ZFS_ROOT = os.environ['ZFS_ROOT']
@@ -39,7 +40,6 @@ def has_certs(website_id, domain, use_www):
     return cert_exists
 
 
-
 def get_cert(website_id, domain, use_www):
     dataset = f'/{ZFS_ROOT}/{website_id}'
     domains = f'-d {domain} -d www.{domain}' if use_www else f'-d {domain}'
@@ -47,8 +47,35 @@ def get_cert(website_id, domain, use_www):
 
 
 def renew_cert(website_id):
+    print('Renewing cert')
+
+    with db.open_db() as conn:
+        with conn.cursor() as cur:
+            query = "SELECT domain,use_https,www FROM users_website WHERE id=%s"
+            cur.execute(query, (website_id,))
+            website_info = cur.fetchone()
+    
+    if not website_info:
+        print('Failed to retrieve website information for site ' + website_id)
+        return
+
+    (domain, use_https, use_www) = website_info
+
+    if not use_https:
+        print('Not renewing certificates, use HTTPS is disabled.')
+        return
+    
+    if not has_certs(website_id, domain, use_www):
+        print('Website does not have certificates yet, attempt to generate them first..')
+        generate_certs(website_id, domain, use_www)
+
+    if not has_certs(website_id, domain, use_www):
+        print('Website still does not have certificates, ABORT!')
+        return
+
     dataset = f'/{ZFS_ROOT}/{website_id}'
     os.system(f'docker run -it --rm -v {dataset}/web:/web -v {dataset}/certs:/etc/letsencrypt certbot/certbot renew')
+    copy_certs(website_id, domain)
 
 
 def copy_certs(website_id, domain):
